fix: change poster style and replaced the gif indicator with a css indicator
