+++ movie-search/src/index.html
@@ -28,7 +28,7 @@
           <button class="icon-button keyboard-button" data-do="keyboard">
           </button>          
         </form>
-        <img class="load-indicator load-indicator_hidden" src="./assets/icons/spin_new.gif" alt="load">        
+        <div class="lds-dual-ring lds-dual-ring_hidden"></div>
       </div>
       <div class="message-container">
         <p class="message">Search results for ...</p>
