+++ movie-search/src/js/control.js
@@ -1,14 +1,14 @@
 import { activateKeyboard, toggleKeyboardVisibility } from './keyboard';
 import { addRequestToSlider, clearSlider, doBeforeSliderEnd } from './slider';
-import { getMovies, getMovieRating, getTranslate } from './network';
+import { sendRequest, getAllMovieWithRating } from './network';
 
 const pageElement = {
   searchContainer: 'search-container',
   searchButton: 'search__button',
   input: 'search__input',
   message: 'message',
-  indicator: 'load-indicator',
-  hideIndicator: 'load-indicator_hidden',
+  indicator: 'lds-dual-ring',
+  hideIndicator: 'lds-dual-ring_hidden',
 };
 
 let currentSearch = 'star';
@@ -22,55 +22,15 @@ const setIndicatorVisible = (isVisible = true) => {
   indicator.classList.add(pageElement.hideIndicator);
 };
 
-const getEnglish = async (request) => {
-  const rusRegexp = /[А-яё]/g;
-  const isNeedTranslate = rusRegexp.test(request);
-  let engRequest = request;
-
-  if (isNeedTranslate) {
-    const yandexResponse = await getTranslate(request);
-    [engRequest] = (yandexResponse.code >= 200 && yandexResponse.code < 300) ? yandexResponse.text : ['translation_error'];
-  }
-
-  return engRequest;
-};
-
 const setMessageText = (text) => {
   const message = document.querySelector(`.${pageElement.message}`);
   message.innerText = text;
 };
 
-const sendRequest = async (request) => {
-  const requestToOmdb = await getEnglish(request);
-  if (requestToOmdb === 'translation_error') {
-    return 'translation_error';
-  }
-  currentSearch = requestToOmdb;
-  const data = await getMovies(requestToOmdb);
-
-  return data;
-};
-
-const getAllMovieWithRating = async (movieList) => {
-  const requestList = movieList.map(async (item) => {
-    const tempClone = { ...item };
-    const { imdbID } = item;
-
-    const response = await getMovieRating(imdbID);
-
-    tempClone.Rating = response.imdbRating;
-    return tempClone;
-  });
-
-  const allResponse = await Promise.allSettled(requestList);
-  const moviesWithRaiting = allResponse.map((item) => item.value);
-
-  return moviesWithRaiting;
-};
-
 const requestAndAdd = async (title, clear = false) => {
   setIndicatorVisible(true);
-  const data = await sendRequest(title);
+  const [data, requestToOmdb] = await sendRequest(title);
+
   if (data === 'translation_error') {
     setMessageText('translation error');
     setIndicatorVisible(false);
@@ -86,7 +46,7 @@ const requestAndAdd = async (title, clear = false) => {
   if (clear) {
     clearSlider();
   }
-
+  currentSearch = requestToOmdb;
   addRequestToSlider(moviesWithRaiting);
   setMessageText(`Search results for ${currentSearch}`);
   setIndicatorVisible(false);
@@ -108,8 +68,7 @@ export const clickHandler = () => {
     if (event.target.dataset.do === 'search') {
       event.preventDefault();
 
-      currentSearch = input.value;
-      await requestAndAdd(currentSearch, true);
+      await requestAndAdd(input.value, true);
       input.focus();
     }
