+++ movie-search/src/sass/base/_indicator.scss
@@ -1,0 +1,29 @@
+.lds-dual-ring {
+  display: inline-block;
+  width: 44px;
+  height: 36px;
+  padding-left: 8px;
+  &_hidden {
+    opacity: 0;
+  }
+}
+
+.lds-dual-ring:after {
+  content: " ";
+  display: block;
+  width: 26px;
+  height: 26px;
+  margin: auto;
+  border-radius: 50%;
+  border: 5px solid $color-primary;
+  border-color: $color-primary transparent $color-primary transparent;
+  animation: lds-dual-ring 1.2s linear infinite;
+}
+@keyframes lds-dual-ring {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
