+++ movie-search/src/sass/components/_card.scss
@@ -11,15 +11,16 @@
 }
 
 .poster-container {
+  width: 100%;
   height: 465px;
-  @extend %flex-centered;
+  @extend %flex-centered;  
   overflow: hidden;
 }
 
 .card__poster {
-  
+  display: block;
   height: 100%;
-  flex-shrink: 0;
+  object-fit: contain;
 }
 
 .card__title {
