+++ movie-search/src/sass/components/_search.scss
@@ -48,14 +48,6 @@
   }  
 }
 
-.load-indicator {
-  width: 36px;
-  padding-left: 3px;
-  &_hidden {
-    opacity: 0;
-  }
-}
-
 @media (max-width: 767px) {
   .search__button-container {
     display: none;
