+++ movie-search/src/sass/index.scss
@@ -3,6 +3,7 @@
 @import "abstract/constants";
 @import "abstract/placeholders";
 @import "base/buttons";
+@import "base/indicator";
 @import "layout/wrapper";
 @import "layout/main";
 @import "layout/swiper";
